Improve publisher test suite grouping
